fix: integrate timeline and graph tabs wrappers under one component (#572)
* fix: integrate timeline and graph tabs wrappers under one component


* fix: details tab in graph view and clean-up


* fix: old unit tests


* fix: temp filename change


* fix: final filename change


* fix: tsconfig
