+++ packages/zapp/console/src/components/Executions/Tables/SelectNodeExecutionLink.tsx
@@ -1,7 +1,6 @@
 import { Link } from '@material-ui/core';
-import { NodeExecution } from 'models/Execution/types';
+import { NodeExecution, NodeExecutionIdentifier } from 'models/Execution/types';
 import * as React from 'react';
-import { NodeExecutionsTableState } from './types';
 
 /** Renders a link that, when clicked, will trigger selection of the
  * given NodeExecution.
@@ -10,14 +9,14 @@ export const SelectNodeExecutionLink: React.FC<{
   className?: string;
   execution: NodeExecution;
   linkText: string;
-  state: NodeExecutionsTableState;
-}> = ({ className, execution, linkText, state }) => {
+  setSelectedExecution: (selectedExecutionId: NodeExecutionIdentifier | null) => void;
+}> = ({ className, execution, linkText, setSelectedExecution }) => {
   // open the side panel for selected execution's detail
   const onClick = (e: React.MouseEvent<HTMLElement>) => {
     // prevent the parent row body onClick event trigger
     e.stopPropagation();
     // use null in case if there is no execution provided - when it is null will close panel
-    state.setSelectedExecution(execution?.id ?? null);
+    setSelectedExecution(execution?.id ?? null);
   };
 
   return (
